+++ SpringRESTPostgreSQL-Server/SpringRESTPostgreSQL/src/main/java/com/javasampleapproach/springrest/postgresql/repo/ManufacturerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.javasampleapproach.springrest.postgresql.model.Manufacturer;
 
-public interface ManufacturerRepository extends CrudRepository<Manufacturer, Integer> {
-	List<Manufacturer> findById(int Id);
+public interface ManufacturerRepository extends CrudRepository<Manufacturer, Long> {
+	List<Manufacturer> findById(long Id);
 }
